os: move `enum ExitCode` into private header
Not used by any external drivers, so no need to keep it public.
