+++ os/ddx_priv.h
@@ -5,7 +5,8 @@
 #ifndef _XSERVER_OS_DDX_PRIV_H
 #define _XSERVER_OS_DDX_PRIV_H
 
-#include "os.h"
+#include "include/os.h"
+#include "os/osdep.h"
 
 /* callbacks of the DDX, which are called by DIX or OS layer.
    DDX's need to implement these in order to handle DDX specific things.
