+++ os/osdep.h
@@ -234,4 +234,11 @@ static inline size_t x_safe_strlen(const char *str) {
     return (str ? strlen(str) : 0);
 }
 
+enum ExitCode {
+    EXIT_NO_ERROR = 0,
+    EXIT_ERR_ABORT = 1,
+    EXIT_ERR_CONFIGURE = 2,
+    EXIT_ERR_DRIVERS = 3,
+};
+
 #endif                          /* _OSDEP_H_ */
